// frontend/src/generated/model/account-create-request-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Contains information that is needed to create an account.
 */
export interface AccountCreateRequestDto { 
    /**
     * Username of the account to be created.
     */
    username: string;
    /**
     * Email of the account to be created.
     */
    email: string;
    /**
     * First name of the user that creates this account.
     */
    firstName: string;
    /**
     * Last name of the user that creates this account.
     */
    lastName: string;
    /**
     * Password of the account to be created.
     */
    password: string;
}